Commit database after every modification
Given the actual workload, this should not result in decrease of
performance or disk wear-out. It, however, simplifies code greatly and
guarantees the latest data is always present on disk, which might help
recover from crashes. It also get rid of a mutex lock in the 'finally'
block, likely eliminating the possibility of deadlocks on shutdown.